+++ packages/next/src/client/components/react-dev-overlay/src/components/code-frame/code-frame.tsx
@@ -5,9 +5,9 @@ import stripAnsi from 'next/dist/compiled/strip-ansi'
 
 import { useMemo } from 'react'
 import { HotlinkedText } from '../hot-linked-text'
-import { getFrameSource } from '../../helpers/stack-frame'
-import { useOpenInEditor } from '../../helpers/use-open-in-editor'
-import { noop as css } from '../../helpers/noop-template'
+import { getFrameSource } from '../../../utils/stack-frame'
+import { useOpenInEditor } from '../../utils/use-open-in-editor'
+import { noop as css } from '../../../utils/noop-template'
 import { ExternalIcon } from '../../icons/external'
 import { FileIcon } from '../../icons/file'
