+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/dev-tools-indicator/dev-tools-indicator.tsx
@@ -7,7 +7,7 @@ import { NextLogo } from './internal/next-logo'
 import { useIsDevBuilding } from '../../../../../../dev/dev-build-indicator/internal/initialize'
 import { useIsDevRendering } from '../../../../utils/dev-indicator/dev-render-indicator'
 import { useDelayedRender } from '../../../hooks/use-delayed-render'
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 import { TurbopackInfo } from './dev-tools-info/turbopack-info'
 import { RouteInfo } from './dev-tools-info/route-info'
 import { StopIcon } from '../../../icons/stop-icon'
