+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/dev-tools-indicator/dev-tools-info/dev-tools-info.tsx
@@ -1,4 +1,4 @@
-import { noop as css } from '../../../../helpers/noop-template'
+import { noop as css } from '../../../../../utils/noop-template'
 
 export function DevToolsInfo({
   title,
