+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/dev-tools-indicator/dev-tools-info/route-info.tsx
@@ -1,5 +1,5 @@
 import { DevToolsInfo } from './dev-tools-info'
-import { noop as css } from '../../../../helpers/noop-template'
+import { noop as css } from '../../../../../utils/noop-template'
 
 function StaticRouteContent() {
   return (
