+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/dev-tools-indicator/internal/next-logo.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, useEffect, useRef, useState } from 'react'
-import { noop as css } from '../../../../helpers/noop-template'
-import mergeRefs from '../../../../helpers/merge-refs'
+import { noop as css } from '../../../../../utils/noop-template'
+import mergeRefs from '../../../../utils/merge-refs'
 import { useMinimumLoadingTimeMultiple } from './use-minimum-loading-time-multiple'
 
 interface Props extends React.ComponentProps<'button'> {
