+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/dialog/body.tsx
@@ -1,4 +1,4 @@
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 import { DialogBody } from '../../dialog'
 
 type ErrorOverlayDialogBodyProps = {
