+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/environment-name-label/environment-name-label.tsx
@@ -1,4 +1,4 @@
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 
 export function EnvironmentNameLabel({
   environmentName,
