+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/error-message/error-message.tsx
@@ -1,5 +1,5 @@
 import { useState, useRef, useLayoutEffect } from 'react'
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 
 export type ErrorMessageType = React.ReactNode
