+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/error-overlay-footer/error-feedback/error-feedback.tsx
@@ -1,8 +1,8 @@
 import { useState, useCallback } from 'react'
 import { ThumbsUp } from '../../../../icons/thumbs/thumbs-up'
 import { ThumbsDown } from '../../../../icons/thumbs/thumbs-down'
-import { noop as css } from '../../../../helpers/noop-template'
-import { cx } from '../../../../helpers/cx'
+import { noop as css } from '../../../../../utils/noop-template'
+import { cx } from '../../../../utils/cx'
 
 interface ErrorFeedbackProps {
   errorCode: string
