+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/error-overlay-layout/error-overlay-layout.tsx
@@ -8,7 +8,7 @@ import {
   styles as toolbarStyles,
 } from '../error-overlay-toolbar/error-overlay-toolbar'
 import { ErrorOverlayFooter } from '../error-overlay-footer/error-overlay-footer'
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 import {
   ErrorMessage,
   styles as errorMessageStyles,
