+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/error-overlay-nav/error-overlay-nav.tsx
@@ -2,7 +2,7 @@ import type { VersionInfo } from '../../../../../../../server/dev/parse-version-
 
 import { ErrorOverlayPagination } from '../error-overlay-pagination/error-overlay-pagination'
 import { VersionStalenessInfo } from '../../version-staleness-info/version-staleness-info'
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 import type { ReadyRuntimeError } from '../../../../utils/get-error-by-type'
 
 type ErrorOverlayNavProps = {
