+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/error-overlay-pagination/error-overlay-pagination.tsx
@@ -5,7 +5,7 @@ import {
   useRef,
   useState,
 } from 'react'
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 import { LeftArrow } from '../../../icons/left-arrow'
 import { RightArrow } from '../../../icons/right-arrow'
 import type { ReadyRuntimeError } from '../../../../utils/get-error-by-type'
