+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/error-overlay-toolbar/docs-link-button.tsx
@@ -1,4 +1,4 @@
-import { parseUrlFromText } from '../../../helpers/parse-url-from-text'
+import { parseUrlFromText } from '../../../utils/parse-url-from-text'
 
 const docsURLAllowlist = ['https://nextjs.org', 'https://react.dev']
