+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/error-overlay-toolbar/error-overlay-toolbar.tsx
@@ -1,6 +1,6 @@
 import type { DebugInfo } from '../../../../types'
 import { NodejsInspectorButton } from './nodejs-inspector-button'
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 import { CopyStackTraceButton } from './copy-stack-trace-button'
 import { DocsLinkButton } from './docs-link-button'
