+++ packages/next/src/client/components/react-dev-overlay/src/components/errors/error-type-label/error-type-label.tsx
@@ -1,4 +1,4 @@
-import { noop as css } from '../../../helpers/noop-template'
+import { noop as css } from '../../../../utils/noop-template'
 
 export type ErrorType =
   | 'Build Error'
