+++ packages/next/src/client/components/react-dev-overlay/src/components/terminal/terminal.tsx
@@ -3,9 +3,9 @@ import * as React from 'react'
 import { HotlinkedText } from '../hot-linked-text'
 import { EditorLink } from './editor-link'
 import { ExternalIcon } from '../../icons/external'
-import { noop as css } from '../../helpers/noop-template'
-import { getFrameSource } from '../../helpers/stack-frame'
-import { useOpenInEditor } from '../../helpers/use-open-in-editor'
+import { noop as css } from '../../../utils/noop-template'
+import { getFrameSource } from '../../../utils/stack-frame'
+import { useOpenInEditor } from '../../utils/use-open-in-editor'
 import { FileIcon } from '../../icons/file'
 
 export type TerminalProps = { content: string }
