+++ packages/next/src/client/components/react-dev-overlay/src/components/toast/styles.ts
@@ -1,4 +1,4 @@
-import { noop as css } from '../../helpers/noop-template'
+import { noop as css } from '../../../utils/noop-template'
 
 const styles = css`
   .nextjs-toast {
