+++ packages/next/src/client/components/react-dev-overlay/src/components/version-staleness-info/version-staleness-info.tsx
@@ -1,6 +1,6 @@
 import type { VersionInfo } from '../../../../../../server/dev/parse-version-info'
-import { cx } from '../../helpers/cx'
-import { noop as css } from '../../helpers/noop-template'
+import { cx } from '../../utils/cx'
+import { noop as css } from '../../../utils/noop-template'
 
 export function VersionStalenessInfo({
   versionInfo,
