+++ packages/next/src/client/components/react-dev-overlay/src/container/errors.tsx
@@ -1,7 +1,7 @@
 import { useState, useMemo, useEffect, useRef, Suspense } from 'react'
 import type { DebugInfo } from '../../types'
 import { Overlay } from '../components/overlay'
-import { noop as css } from '../helpers/noop-template'
+import { noop as css } from '../../utils/noop-template'
 import { RuntimeError } from './runtime-error'
 import { getErrorSource } from '../../../../../shared/lib/error-source'
 import { HotlinkedText } from '../components/hot-linked-text'
