+++ packages/next/src/client/components/react-dev-overlay/src/container/runtime-error/component-stack-pseudo-html.tsx
@@ -1,4 +1,4 @@
-import { noop as css } from '../../helpers/noop-template'
+import { noop as css } from '../../../utils/noop-template'
 
 export { PseudoHtmlDiff } from '../../components/hydration-diff/diff-view'
