--- packages/next/src/client/components/react-dev-overlay/src/helpers/cx.ts
@@ -1,9 +1,0 @@
-/**
- * Merge multiple args to a single string with spaces. Useful for merging class names.
- * @example
- * cx('foo', 'bar') // 'foo bar'
- * cx('foo', null, 'bar', undefined, 'baz', false) // 'foo bar baz'
- */
-export function cx(...args: (string | undefined | null | false)[]): string {
-  return args.filter(Boolean).join(' ')
-}
