+++ packages/next/src/client/components/react-dev-overlay/src/styles/colors.tsx
@@ -1,4 +1,4 @@
-import { noop as css } from '../helpers/noop-template'
+import { noop as css } from '../../utils/noop-template'
 
 // TODO: Replace the existing colors in Base.tsx.
 export function Colors() {
