+++ packages/next/src/client/components/react-dev-overlay/src/styles/css-reset.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { noop as css } from '../helpers/noop-template'
+import { noop as css } from '../../utils/noop-template'
 
 export function CssReset() {
   return (
