+++ packages/next/src/client/components/react-dev-overlay/src/utils/parse-url-from-text.ts
