+++ packages/next/src/client/components/react-dev-overlay/src/utils/use-open-in-editor.ts
